requirements.txt
streamlit
pdfminer.six
sentence-transformers
scikit-learn
groq
python-dotenv
nltk
pandas
plotly
